requirements.txt
langchain
langchain-openai
langchain-community
langgraph
chromadb
python-dotenv
pypdf
python-docx
reportlab
tiktoken
fastapi
uvicorn
jinja2
python-multipart
python-jose[cryptography]
bcrypt
sqlalchemy
aiofiles
httpx
Pillow
pdf2image
pymupdf